wip: this is incomplete refactor
