+++ include/sddf/util/vspace.h
@@ -1,0 +1,3 @@
+#include <microkit.h>
+
+bool map_memory_region(uintptr_t paddr, uintptr_t size, uintptr_t vaddr);
